split out table into seperate components
