+++ web-app-router/scenes/Portal/Teams/TeamId/Apps/AppId/Actions/page/ActionsList/ActionsList.tsx
@@ -1,7 +1,10 @@
 "use client";
 import { Button } from "@/components/Button";
-import { TableComponent } from "@/components/Table";
-import { useState } from "react";
+import { Table } from "@/components/Table";
+import { Body } from "@/components/Table/Body";
+import { Footer } from "@/components/Table/Footer";
+import { Header } from "@/components/Table/Header";
+import { useMemo, useState } from "react";
 
 // TODO: Example of how to use this component
 export const ActionsList = () => {
@@ -20,7 +23,7 @@ export const ActionsList = () => {
       </div>,
       <span key={1}>30</span>,
       <span key={1}>New York</span>,
-      <Button type="button" key={1} className="rotate-90">
+      <Button type="button" key={1} className="rotate-90 w-full">
         ...
       </Button>,
     ],
@@ -31,35 +34,53 @@ export const ActionsList = () => {
       </div>,
       <span key={2}>25</span>,
       <span key={2}>London</span>,
-      <Button type="button" key={2} className="rotate-90">
+      <Button type="button" key={2} className="rotate-90 w-full">
         ...
       </Button>,
     ],
+
     // ... more rows
   ];
 
-  const [totalResults, setTotalResults] = useState(2); // Total number of results
+  const [totalResults, setTotalResults] = useState(rows.length); // Total number of results
+  const [currentPage, setCurrentPage] = useState(1);
+  const [rowsPerPage, setRowsPerPage] = useState(10); // Rows per page
   const rowsPerPageOptions = [10, 20, 50]; // Rows per page options
 
   const handlePageChange = (page: number) => {
     // Logic to handle page change
+    setCurrentPage(page);
   };
 
   const handleRowsPerPageChange = (rowsPerPage: number) => {
     // Logic to handle rows per page change
+    setRowsPerPage(rowsPerPage);
+    setCurrentPage(1); // Reset to first page when rows per page changes
   };
+
+  const setRowsToDisplay = useMemo(() => {
+    const startIndex = (currentPage - 1) * rowsPerPage;
+    const endIndex = startIndex + rowsPerPage;
+    return rows.slice(startIndex, endIndex);
+  }, [currentPage, rowsPerPage, rows]);
 
   return (
     <div className="flex items-center justify-center w-[700px]">
-      <TableComponent
-        headers={headers}
-        rows={rows}
-        totalResults={totalResults}
-        rowsPerPageOptions={rowsPerPageOptions}
-        onPageChange={handlePageChange}
-        onRowsPerPageChange={handleRowsPerPageChange}
-        currentPage={1}
-      />
+      <Table
+        footer={
+          <Footer
+            totalResults={totalResults}
+            currentPage={currentPage}
+            rowsPerPage={rowsPerPage}
+            rowsPerPageOptions={rowsPerPageOptions}
+            handlePageChange={handlePageChange}
+            handleRowsPerPageChange={handleRowsPerPageChange}
+          />
+        }
+      >
+        <Header headers={headers} />
+        <Body rows={setRowsToDisplay} />
+      </Table>
     </div>
   );
 };
